* DROOLS-3826: Implementing review commentaries
