+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/main/java/org/kie/workbench/common/dmn/client/graph/DMNGraphUtils.java
@@ -17,20 +17,16 @@
 package org.kie.workbench.common.dmn.client.graph;
 
 import java.util.Optional;
-import java.util.stream.StreamSupport;
 
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
 
-import org.kie.workbench.common.dmn.api.definition.v1_1.DMNDiagram;
 import org.kie.workbench.common.dmn.api.definition.v1_1.Definitions;
+import org.kie.workbench.common.dmn.api.graph.DMNDiagramUtils;
 import org.kie.workbench.common.stunner.core.client.api.SessionManager;
 import org.kie.workbench.common.stunner.core.client.canvas.CanvasHandler;
 import org.kie.workbench.common.stunner.core.client.session.ClientSession;
 import org.kie.workbench.common.stunner.core.diagram.Diagram;
-import org.kie.workbench.common.stunner.core.graph.Graph;
-import org.kie.workbench.common.stunner.core.graph.Node;
-import org.kie.workbench.common.stunner.core.graph.content.definition.Definition;
 
 @Dependent
 public class DMNGraphUtils {
@@ -39,41 +35,31 @@ public class DMNGraphUtils {
 
     private SessionManager sessionManager;
 
+    private DMNDiagramUtils dmnDiagramUtils;
+
     public DMNGraphUtils() {
         //CDI proxy
     }
 
     @Inject
-    public DMNGraphUtils(final SessionManager sessionManager) {
+    public DMNGraphUtils(final SessionManager sessionManager,
+                         final DMNDiagramUtils dmnDiagramUtils) {
         this.sessionManager = sessionManager;
+        this.dmnDiagramUtils = dmnDiagramUtils;
     }
 
     public Definitions getDefinitions() {
         return getCurrentSession()
                 .map(clientSession -> {
                     return getCanvasHandler(clientSession)
-                            .map(canvasHandler -> getDefinitions(canvasHandler.getDiagram()))
+                            .map(canvasHandler -> dmnDiagramUtils.getDefinitions(canvasHandler.getDiagram()))
                             .orElse(NONE);
                 })
                 .orElse(NONE);
     }
 
-    @SuppressWarnings("unchecked")
     public Definitions getDefinitions(final Diagram diagram) {
-
-        final Graph<?, Node> graph = diagram.getGraph();
-
-        return StreamSupport
-                .stream(graph.nodes().spliterator(), false)
-                .map(Node::getContent)
-                .filter(c -> c instanceof Definition)
-                .map(c -> (Definition) c)
-                .map(Definition::getDefinition)
-                .filter(d -> d instanceof DMNDiagram)
-                .map(d -> (DMNDiagram) d)
-                .findFirst()
-                .map(DMNDiagram::getDefinitions)
-                .orElse(NONE);
+        return dmnDiagramUtils.getDefinitions(diagram);
     }
 
     private Optional<ClientSession> getCurrentSession() {
