+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/test/java/org/kie/workbench/common/dmn/client/graph/DMNGraphUtilsTest.java
@@ -19,19 +19,18 @@ package org.kie.workbench.common.dmn.client.graph;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kie.workbench.common.dmn.api.definition.v1_1.DMNDiagram;
-import org.kie.workbench.common.dmn.api.definition.v1_1.Decision;
 import org.kie.workbench.common.dmn.api.definition.v1_1.Definitions;
+import org.kie.workbench.common.dmn.api.graph.DMNDiagramUtils;
 import org.kie.workbench.common.stunner.core.client.api.SessionManager;
 import org.kie.workbench.common.stunner.core.client.canvas.CanvasHandler;
 import org.kie.workbench.common.stunner.core.client.session.ClientSession;
+import org.kie.workbench.common.stunner.core.diagram.Diagram;
 import org.kie.workbench.common.stunner.core.graph.Edge;
 import org.kie.workbench.common.stunner.core.graph.Node;
 import org.kie.workbench.common.stunner.core.graph.content.Bounds;
 import org.kie.workbench.common.stunner.core.graph.content.definition.DefinitionSet;
 import org.kie.workbench.common.stunner.core.graph.content.view.View;
 import org.kie.workbench.common.stunner.core.graph.content.view.ViewImpl;
-import org.kie.workbench.common.stunner.core.graph.impl.EdgeImpl;
 import org.kie.workbench.common.stunner.core.graph.impl.GraphImpl;
 import org.kie.workbench.common.stunner.core.graph.impl.NodeImpl;
 import org.kie.workbench.common.stunner.core.graph.store.GraphNodeStoreImpl;
@@ -44,15 +43,17 @@ import org.mockito.runners.MockitoJUnitRunner;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class DMNGraphUtilsTest {
 
-    private static final String NAME = "name";
-
     @Mock
     private SessionManager sessionManager;
+
+    @Mock
+    private DMNDiagramUtils dmnDiagramUtils;
 
     @Mock
     private ClientSession clientSession;
@@ -61,80 +62,45 @@ public class DMNGraphUtilsTest {
     private CanvasHandler canvasHandler;
 
     @Mock
-    private ProjectMetadata metadata;
-
-    @Mock
-    private Bounds bounds;
+    private Diagram diagram;
 
     private DMNGraphUtils utils;
 
-    private GraphImpl<DefinitionSet> graph;
-
     @Before
     public void setup() {
-        this.utils = new DMNGraphUtils(sessionManager);
-        this.graph = new GraphImpl<>(UUID.uuid(), new GraphNodeStoreImpl());
-        final ProjectDiagramImpl diagram = new ProjectDiagramImpl(NAME, graph, metadata);
+
+        utils = new DMNGraphUtils(sessionManager, dmnDiagramUtils);
+
+
         when(sessionManager.getCurrentSession()).thenReturn(clientSession);
         when(clientSession.getCanvasHandler()).thenReturn(canvasHandler);
         when(canvasHandler.getDiagram()).thenReturn(diagram);
     }
 
     @Test
-    public void testGetDefinitionsWithRootNode() {
-        final DMNDiagram definition = new DMNDiagram();
-        graph.addNode(newNode(definition));
+    public void testGetDefinitions() {
 
-        final Definitions definitions = utils.getDefinitions();
-        assertNotNull(definitions);
-        assertEquals(definition.getDefinitions(),
-                     definitions);
+        final Definitions expectedDefinitions = mock(Definitions.class);
+
+        when(dmnDiagramUtils.getDefinitions(diagram)).thenReturn(expectedDefinitions);
+
+        final Definitions actualDefinitions = utils.getDefinitions();
+
+        assertNotNull(actualDefinitions);
+        assertEquals(expectedDefinitions, actualDefinitions);
     }
 
     @Test
-    public void testGetDefinitionsWithMultipleRootNodes() {
-        final Decision definition1 = new Decision();
-        final DMNDiagram definition2 = new DMNDiagram();
-        graph.addNode(newNode(definition1));
-        graph.addNode(newNode(definition2));
+    public void testGetDefinitionsWithDiagram() {
 
-        final Definitions definitions = utils.getDefinitions();
-        assertNotNull(definitions);
-        assertEquals(definition2.getDefinitions(),
-                     definitions);
-    }
+        final Definitions expectedDefinitions = mock(Definitions.class);
+        final Diagram diagram = mock(Diagram.class);
 
-    @Test
-    public void testGetDefinitionsWithConnectedNodes() {
-        final Decision definition1 = new Decision();
-        final DMNDiagram definition2 = new DMNDiagram();
-        final Node<View, Edge> node1 = newNode(definition1);
-        final Node<View, Edge> node2 = newNode(definition2);
+        when(dmnDiagramUtils.getDefinitions(diagram)).thenReturn(expectedDefinitions);
 
-        final Edge<View, Node> edge = new EdgeImpl<>(UUID.uuid());
-        node1.getInEdges().add(edge);
-        node2.getOutEdges().add(edge);
-        edge.setSourceNode(node2);
-        edge.setTargetNode(node1);
+        final Definitions actualDefinitions = utils.getDefinitions(diagram);
 
-        graph.addNode(node1);
-        graph.addNode(node2);
-
-        final Definitions definitions = utils.getDefinitions();
-        assertNotNull(definitions);
-        assertEquals(definition2.getDefinitions(),
-                     definitions);
-    }
-
-    @Test
-    public void testGetDefinitionsWithNoNodes() {
-        assertNull(utils.getDefinitions());
-    }
-
-    private Node<View, Edge> newNode(final Object definition) {
-        final Node<View, Edge> node = new NodeImpl<>(UUID.uuid());
-        final View<Object> content = new ViewImpl<>(definition, bounds);
-        node.setContent(content);
-        return node;
+        assertNotNull(actualDefinitions);
+        assertEquals(expectedDefinitions, actualDefinitions);
     }
 }
